Added threads; check description
vApplicationMallocFailedHook() will only be called if
configUSE_MALLOC_FAILED_HOOK is set to 1 in FreeRTOSConfig.h.  It is a
hook
function that will get called if a call to pvPortMalloc() fails.
pvPortMalloc() is called internally by the kernel whenever a task,
queue,
timer or semaphore is created.  It is also called by various parts of
the
demo application.

GETTING THIS ERROR