Fixed bug!
Found why the registers were not working and the tests failed.
It was because it read/write to the wrong register CCR instead of the correct ones SR and CMR. Also the template was updated by guidance from
Mathias but there is no confirmation that this was part of the problem.