All tests working on TC
The tests are running at TC0 - channel 0, TC1 - channel 1 and TC2 - channel 0. This is to test that all three base adresses and all the offsets are working even if it is a different channel.

Also the delays are working. One fast delay, one a bit slower and one really slow.